move modules to different directory
